fix: add ingest pipelines plugins template
This has gone unnoticed in the previous version because the common-elasticsearch version was not bumped.

This revision adds the pipeline templates and bumpts common-elasticsearch as well as reporter-common in
order to get the freemarker component able to read templates from a parent class loader.

see https://github.com/gravitee-io/gravitee-reporter-common/pull/4